backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
PyMySQL==1.1.0
bcrypt==4.1.1
python-dotenv==1.0.0
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.26.1
Werkzeug==2.3.7
google-generativeai==0.5.4
reportlab==4.0.7
pyotp==2.9.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78